Increase timeout for reboot

+++ tests/make-backup-and-restore-iso/runtest.sh
@@ -135,8 +135,12 @@ initrd16 (\$rootfs)$ROOT_PATH/$REAR_ISO_OUTPUT/rear-rescue-only.iso
 set default=\"ReaR-recover\"' >> /boot/grub2/grub.cfg" 0 "Setup GRUB"
         rlPhaseEnd
 
-        rhts-reboot
-
+        if test "$TMT_REBOOT_COUNT"; then
+            rlRun "tmt-reboot -t 1200" 0 "Reboot the machine"
+        else
+            # not running from TMT
+            rhts-reboot
+        fi
     elif [ "$REBOOTCOUNT" -eq 1 ]; then
         # ReaR hopefully recovered the OS
         rlPhaseStartTest "Assert that the recovery was successful"
